+++ src/services/editor/keyListener.js
@@ -3,7 +3,12 @@ import keyListenerSvc from '../keyListenerSvc';
 import KeyStrokes from '../keystrokes';
 import store from '../../store';
 import config from '../../config';
-import { ADT_LINK_SELECTOR, CODING_SELECTOR, ZERO_WIDTH } from '../../config/constants';
+import {
+  ADT_LINK_SELECTOR,
+  CODING_SELECTOR,
+  ICON_SELECTOR,
+  ZERO_WIDTH
+} from '../../config/constants';
 import selectionSvc, { RangeCursor } from '../selectionSvc';
 import utils from '../utils';
 
@@ -92,41 +97,32 @@ export default {
   _escapeElement(evt, escapeAfter = true) {
     let $codingArea = null;
     let $adtLinkArea = null;
-    let isCoding = false;
-    let isBlockQuote = false;
-    let isADTLink = false;
 
     const { range, containerElement, sel: selection } = selectionSvc.getSelection();
     const $nearestBlockquote = $(containerElement).closest('blockquote, .alert');
 
+    // is it a blockquote?
     if ($nearestBlockquote.length) {
-      isBlockQuote = true;
+      insertBreak($nearestBlockquote[0], range, escapeAfter);
+      updateRange(selection, range, evt);
     } else {
+      // is it a coding block?
       $codingArea = $(containerElement).closest(CODING_SELECTOR);
       if ($codingArea.length) {
-        isCoding = true;
+        if ($codingArea.prop('tagName') === 'DIV') {
+          insertBreak($codingArea[0], range, escapeAfter);
+        } else {
+          insertSpace($codingArea[0], range, escapeAfter);
+        }
+        updateRange(selection, range, evt);
       } else {
         // is it inside ADT Link?
         $adtLinkArea = $(containerElement).closest(ADT_LINK_SELECTOR);
         if ($adtLinkArea.length) {
-          isADTLink = true;
+          insertSpace($adtLinkArea[0], range, escapeAfter);
+          updateRange(selection, range, evt);
         }
       }
-    }
-
-    if (isCoding) {
-      if ($codingArea.prop('tagName') === 'DIV') {
-        insertBreak($codingArea[0], range, escapeAfter);
-      } else {
-        insertSpace($codingArea[0], range, escapeAfter);
-      }
-      updateRange(selection, range, evt);
-    } else if (isBlockQuote) {
-      insertBreak($nearestBlockquote[0], range, escapeAfter);
-      updateRange(selection, range, evt);
-    } else if (isADTLink) {
-      insertSpace($adtLinkArea[0], range, escapeAfter);
-      updateRange(selection, range, evt);
     }
   },
   _createEscapeListeners() {
@@ -307,9 +303,12 @@ export default {
     keyListenerSvc.addKeyListener(KeyStrokes.Backspace, { alt: true }, evt => {
       const { containerElement } = selectionSvc.getSelection();
 
-      const $codingArea = $(containerElement).closest(CODING_SELECTOR);
-      if ($codingArea.length) {
-        $codingArea.remove();
+      let $elementToRemove = $(containerElement).closest(CODING_SELECTOR);
+      if (!$elementToRemove.length) {
+        $elementToRemove = $(containerElement).closest(ICON_SELECTOR);
+      }
+      if ($elementToRemove.length) {
+        $elementToRemove.remove();
         evt.preventDefault();
       }
     });